requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
feedparser>=6.0.0
gspread>=6.0.0
